+++ packages/orm-integration/src/various.ts
@@ -1,5 +1,17 @@
 import { expect } from '@jest/globals';
-import { AutoIncrement, BackReference, cast, entity, isReferenceInstance, PrimaryKey, Reference, Unique, uuid, UUID } from '@deepkit/type';
+import {
+    AutoIncrement,
+    BackReference,
+    cast,
+    DatabaseField,
+    entity,
+    isReferenceInstance,
+    PrimaryKey,
+    Reference,
+    Unique,
+    uuid,
+    UUID,
+} from '@deepkit/type';
 import { identifier, sql, SQLDatabaseAdapter } from '@deepkit/sql';
 import { DatabaseFactory } from './test.js';
 import { hydrateEntity, isDatabaseOf, UniqueConstraintFailure } from '@deepkit/orm';
@@ -8,6 +20,44 @@ import { randomBytes } from 'crypto';
 Error.stackTraceLimit = 20;
 
 export const variousTests = {
+    async testOneToOneCircularReferenceRelation(databaseFactory: DatabaseFactory) {
+        @entity.name('totocrr_inventory')
+        class Inventory {
+            id: number & PrimaryKey & AutoIncrement = 0;
+            constructor(public user: User & Reference) {}
+        }
+
+        @entity.name('totocrr_user')
+        class User {
+            id: number & PrimaryKey & AutoIncrement = 0;
+            inventory: Inventory & BackReference = new Inventory(this);
+        }
+
+        const database = await databaseFactory([Inventory, User]);
+
+        const user = cast<User>({});
+
+        await database.persist(user.inventory, user);
+    },
+    async testSkipDatabaseFieldForInserts(databaseFactory: DatabaseFactory) {
+        @entity.name('test_skip_database_field_insert')
+        class User {
+            id: number & PrimaryKey & AutoIncrement = 0;
+            username?: string & DatabaseField<{ skip: true }>
+        }
+
+        const database = await databaseFactory([User]);
+
+        const user = cast<User>({ username: 'peter' });
+
+        await database.persist(user);
+
+        {
+            const result = await database.query(User).findOne();
+            expect(result.id).toBe(1)
+            expect(result).not.toHaveProperty('username');
+        }
+    },
     async testRawQuery(databaseFactory: DatabaseFactory) {
         @entity.name('test_connection_user')
         class user {
@@ -626,6 +676,55 @@ export const variousTests = {
             expect(products[0].product === products[1].product).toBe(true);
             expect(products[0].product.images!.length).toBe(2);
             expect(products[1].product.images!.length).toBe(2);
+        }
+    },
+    async nestedEmbeddedTypes(databaseFactory: DatabaseFactory) {
+        class Country {
+            name!: string;
+            id!: string;
+        }
+        class Author {
+            name!: string;
+            country: Country = new Country();
+        }
+
+        @entity.collection('nested-entity')
+        class Book {
+            id!: number & PrimaryKey;
+            name!: string;
+            author: Author = new Author();
+            anyType: any = {};
+        }
+        const database = await databaseFactory([Book]);
+
+        {
+            const book = new Book();
+            book.id = 12345678;
+            book.name = 'Book1';
+            book.author.name = 'Author1';
+            book.author.country.name = 'Country1';
+            book.author.country.id = '1';
+            book.anyType.test1 = '1';
+            book.anyType.test2 = {deep1: '1'};
+            await database.persist(book);
+        }
+
+        {
+            const session = database.createSession();
+            const book = await session.query(Book).filter({ id: 12345678 }).findOne();
+            book.author.country.name = 'somewhere';
+            book.anyType.test1 = '2';
+            book.anyType.test2.deep1 = '2';
+            await session.commit();
+            console.log(book);
+        }
+
+        {
+            const session = database.createSession();
+            const book = await session.query(Book).filter({ id: 12345678 }).findOne();
+            expect(book.author.country.name).toBe('somewhere');
+            expect(book.anyType.test1).toBe('2');
+            expect(book.anyType.test2.deep1).toBe('2');
         }
     }
 };